+++ src/Symfony/Component/Messenger/Bridge/Redis/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Implement the `KeepaliveReceiverInterface` to enable asynchronously notifying Redis that the job is still being processed, in order to avoid timeouts
+ * Add support for `valkey:` / `valkeys:` schemes
 
 6.3
 ---